Add a JavaScript view for showing modal dialogs.
RB.DialogView is a wrapper around $.modalBox() that makes it easy to
construct, show, and hide modal dialogs. It can be instantiated and
bound to an element to show the contents of that element as a dialog, or
it can be subclassed to allow another view to act as a dialog.

Reviewed at https://reviews.reviewboard.org/r/6305/