Add better security and setup for incoming Google Code web hooks.
The Google Code close-submitted webhook now requires a hooks_uuid value,
which will be validated against the repository, returning a 404 if it
doesn't match.

There's some instructions for configuring the hook now as well, which
will direct the user to the proper page and provide them the URL to
configure.

To make this work, I had to fix a bug in our repository configuration
support, which didn't associate a HostingServiceAccount unless that
account required authorization. It attempted to build the account, but
never saved it. This was a simple fix.

Reviewed at https://reviews.reviewboard.org/r/6316/