Fix issues with the new GitHub post-commit hook handler.
The new GitHub post-commit hook handler had some issues in production.
First, when adding the hook, GitHub would try to send a "ping" event,
which we did not support. This resulted in GitHub saying that the hook
wasn't functioning properly (though it wouldn't disable the hook).

Second, and more importantly, the hook signature verification failed.
The X-Hub-Signature header contained the prefix "sha1=", which the
handler wasn't expecting. It now looks for this, erroring out if it
doesn't find it.

Reviewed at https://reviews.reviewboard.org/r/6318/