Fix another SVN unicode problem.
When there's a file-not-found error with unicode paths in SVN, we were trying
to format the encoded string into a unicode string to create the error message.
This caused a UnicodeDecodeError when it tried to coerce it using the 'ascii'
codec. I've changed the `FileNotFoundError.__init__` method to cope with paths
which are bytes as well as paths which are unicode.

Testing done: Ran unit tests.

Fixes bug 3425.

Reviewed at https://reviews.reviewboard.org/r/6326/