Don't crash the dashboard if a bug can't be linked.
In the case where the bugs field contains characters which aren't valid in the
URL, creating the link would fail with NoReverseMatch. In the case discovered,
a URL was entered in the field instead of a bug number.

This change fixes it so that if any URLs don't reverse, we just stick the text
in without a link.

Testing done:
- Set up a reproduction case. Before the change, my dashboard gives a
  traceback. After, I see my dashboard correctly.
- Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/6329/