Fix sample usage instructions for post-commit hooks on Local Sites.
The usage instructions were showing review request URLs off the root of
the site, rather than ones from a Local Site. We now generate a URL (and
a sample ID, for consistency) and send it off to the template for
rendering.

There's also a fix for some non-localized usage text ("Or:").

Reviewed at https://reviews.reviewboard.org/r/6353/