Fix KeyError: u'message' in SVN commits backend.
Not every commit we get has a message, which can cause exceptions to be thrown
from the repository commits resource. We already had one fix for this sort of
error with the author fields. This change fixes it up to do the same for
message and author everywhere.

Testing done:
Ran unit tests.

Addresses part of bug 3579.

Reviewed at https://reviews.reviewboard.org/r/6361/