Normalize the hooks_uuid field before saving the repository.
The hooks_uuid field would become a blank string after editing in the
admin UI. This resulted in a collision as soon as there was more than
one entry with a blank string.

We now normalize this to None if blank prior to saving.

Reviewed at https://reviews.reviewboard.org/r/6362/