Bullet-proof pre-commit UI a bit.
When receiving an unexpected error from the server, the frontend code for the
pre-commit upload would hit a TypeError, preventing it from completing the
process and leaving the browser with a forever "Loading..." message. This
change updates it to check when the response can't be parsed as JSON and show
"Unknown error" instead.

While I was in here I fixed up some of the nearby localization to be more
correct.

Testing done:
- Checked that uploading a diff through the pre-commit UI still worked as
  expected.
- Ran js-tests.

Reviewed at https://reviews.reviewboard.org/r/6365/