Add documentation on creating the database.
This adds a section at the beginning of the "Creating a Review Board Site"
document which shows the steps for creating a new database and a user to access
it for both MySQL and PostgreSQL. For MySQL, this makes sure that it shows how
to turn on UTF-8 before doing anything.

Testing done:
Built HTML docs.

Fixes bug 3000.

Reviewed at https://reviews.reviewboard.org/r/6377/