Post-commit hooks now match repositories along with commit IDs.
When looking up review requests for a commit, we'd attempt to find one
that has a matching commit_id field if the commit description didn't
explicitly specify an ID. However, this lookup did not take the
repository into account, resulting in problems if two review requests
across two different repositories had the same commit ID.

This is not a major problem for Git and Mercurial. In theory, it could
affect Subversion on Google Code, but we actually have that matching
disabled (though we likely suffered from another bug as a result).

The repository is now passed to the matcher function, and is required
for commit ID matching.

Reviewed at https://reviews.reviewboard.org/r/6382/