Fix issue counter initialization with unopened issues containing states.
It's possible for a Comment to have both issue_opened=False and
issue_status=OPEN. We weren't checking for this during counter
initialization, resulting in such comments being included in the count.

The reason why they end up in this state is a separate issue that should
be addressed, but it's certainly easy to work around. We now properly
check the issue_opened field before even considering its issue_status.

There's also a fix for transitioning an issue status when saving the
Comment when issue_opened is False. This is needed for tests, primarily,
but would probably also solve some issues that could come up in the API.

Reviewed at https://reviews.reviewboard.org/r/6384/