Switch to a different version of marked.js that fixes in-word underscores.
marked.js has an infuriating bug that even when "github flavored markdown" is
turned on, it has problems with underscores in the middle of words. Worse, it
didn't handle escaped underscores in the middle of hyperlinks, so even when
they were escaped, it was broken in some cases.

This change pulls in a version of marked.js from an open pull request that
fixes these bugs, which means that we no longer have to escape them.

Testing done:
- Ran unit tests
- Posted some changes with underscores mid-word, saw that the escaping was gone
  and that all text looked correct (with no unexpected em or strong).

Reviewed at https://reviews.reviewboard.org/r/6415/