Turn off guessing of languages in Markdown code blocks.
We previously used Python-Markdown's default behavior of guessing
the language based on the contents of the code. This was often wrong, or
didn't result in anything meaningful. Fortunately, that almost never
came up in usage (since we were using marked.js and
google-code-highlight for the rendering), but when we move to
server-side-only rendering of Markdown content, it'll become more
important.

Now, syntax highlighting is only enabled if using fenced code blocks
with a language specified.

The manual has been updated for these changes.

Reviewed at https://reviews.reviewboard.org/r/6422/