Use the new close_description field when working with close descriptions.
Now that we have a new, improved field for specifying close descriptions
on a review request, it's time to make use of it in the UI, along with
cleaning up the code surrounding it.

The code dealing with "changeDescription" has been split up a bit to
work with either "changeDescription" or "closeDescription," and to look
in the ReviewRequest or DraftReviewRequest, depending on the field.

This also helps set up for bits of the new server-side Markdown change.

Reviewed at https://reviews.reviewboard.org/r/6424/