Enable per-user defaults for Markdown.
This adds a setting in the My Account page for controlling whether
Markdown is enabled by default. When enabled, all text fields will
default to Markdown mode. When disabled, only fields already set for
rich text in the database will be in Markdown mode by default.

The default for users will be based on the 'default_use_rich_text'
setting in siteconfig. Right now, this is not exposed in the UI, but it
will be in a separate change.

There were a few places left in the codebase where we were assigning
default values for rich_text fields, and those have been removed,
allowing the UI to control the default based on this setting instead.

Reviewed at https://reviews.reviewboard.org/r/6517/