Give file attachments an "Actions" menu.
This change moves the "Add Comment" or "Review" items into an "Actions"
drop-down within the file attachment box. This is in preparation for adding an
"Update" action which will allow the user to replace an existing attachment
with a new version.

In order to make the drop-down not super ugly, I've changed the bottom-left
corner of the "Actions" label to be square instead of rounded.

Testing done:
- Verified the visual appearance of the different menus, both as-is and with
  the "Update" button un-commented.
- Checked that the actions still worked.

Reviewed at https://reviews.reviewboard.org/r/6352/