Fix file storage admin page not saving (always reverts back to local storage option)
Saving the file attachment storage in the admin would fail, reverting back to
the local storage mechanism. This was because the form was trying to validate
itself but it couldn't because OpenStack Swift fields required no matter which
storage option was chosen. The fix is to set OpenStack Swift fields to be not
required when selecting a storage option that is not OpenStack Swift.

Testing Done:
Tested in Chrome, Firefox and IE:
- Hit save when local storage is chosen -> local storage is selected on refresh.
- Amazon S3 filled out and saved -> Change to OpenStack Swift, leave everything
  blank -> Save -> Navigate away and come back -> Amazon S3 chosen. (Reverse
  Amazon S3 and OpenStack Swift works too).
- Amazon S3 (or OpenStack Swift) filled out and saved -> Change to local
  storage -> Save -> Navigate away and come back -> Local storage chosen.
- Local storage saved -> Change to Amazon S3 (or OpenStack Swift), leave
  everything blank -> Navigate away and come back -> Local storage chosen.
- File attachment uploads to Amazon S3 and local storage works. Unable to test
  with OpenStack Swift since no account available to use.

Reviewed at https://reviews.reviewboard.org/r/6432/