Update the new review dialog for server-side Markdown.
The new server-side Markdown support was implemented on release-2.0.x,
and thus didn't take into account the new review dialog on master. This
dialog differed from the other places where Markdown was used, in that
we had no way to pre-inject the raw values for review and comment text
fields prior to loading the dialog.

Now, the dialog requests the raw text fields as part of the initial
fetches and the saves, and populates the raw values in the
inlineEditors. This allows the dialog to work as it did before, with the
addition of having server-side Markdown support.

Reviewed at https://reviews.reviewboard.org/r/6438/