Preserve query args when redirecting from Review/Reply draft resources.
The Review/Reply Draft resources redirect to a resource with the
appropriate ID, but when doing so, they don't preserve any query string
arguments. This means that if something like ?api_format or
?force-text-type is used, it will be lost.

We now have a useful utility function for building a URL that preserves
all query string arguments that aren't automatically preserved by the
base WebAPIResource. These are used in all two resources that do a
redirect from a GET.

Reviewed at https://reviews.reviewboard.org/r/6437/