Fix a regression in the review request API.
The recent change to the close description fields to handle markdown caused a
regression in the case where the close_description field was null. In this
case, trying to normalize the text would fail, and the API endpoint would
return a 500 with a traceback.

This change makes it so that if a text field is null, we just pass that
through.

Testing done:
- Verified that I could once again update un-published review request drafts.
- Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/6441/