Prevent various inline editors from vertically shifting on open/close.
The inline editors in various places (replies to reviews/comments and in
the new review dialog) would shift a bit vertically when opening or
closing the editor. This was due to the pencil disappearing and the line
height shrinking where it used to be.

We had this solved for banners. Now, that solution is contained in a
macro, which is used in the other cases where we display these.

Reviewed at https://reviews.reviewboard.org/r/6446/