Allow fine-grained setting of rich text modes per-field.
When we introduced Markdown support, we added a single rich_text field
for each supported model, and all fields were tied to that setting.
However, if a person wanted to use Markdown for one field but not
another, they were out of luck. That also required that we perform some
complex escaping/unescaping of all fields on a model when switching text
modes.

This is the first change in a series that splits the rich_text field,
providing per-field flags. The changes are all inter-dependent, but
split up based on layers of the codebase.

This change adds the new rich_text fields, updates logic and templates
to point to the new fields, and removes usage of the old fields.

Reviewed at https://reviews.reviewboard.org/r/6456/