Fix a regression in saving comments.
The new delayed notifications for "change" events from MarkdownEditor
could result in text being chopped off if typing and saving within half
a second. We now make sure to always set the latest-and-greatest text
when saving, instead of relying on the handler for the "change" event.

This only occurs in the comment dialog.

Reviewed at https://reviews.reviewboard.org/r/6457/