Backport some of the review dialog Markdown changes from 2.1.
This backports much of the changes made to the review dialog in 2.1 for
Markdown support. Many of the changes made are necessary in 2.0.x as
well.

Primarily, this gives us support for including raw text fields, fetching
comments with a forced text type, and specifying extra query arguments
in drafts.