Add support for raw text fields and text types in JS resource models.
Several JS resource models now have support for fetching raw text fields
and storing them as attributes on the model. This will be used by
Markdown editors to load the raw text for editing while using HTML or
Markdown-escaped for display.

The models can now have includeRawTextFields set to trigger the storage
of the raw fields, which will be stored as the rawTextFields attribute.

Along with this, we now always set the richText fields based on the raw
values, and not the forced values, since we only care about the former.

On top of this, there's also a bug fix for allowing the values in
extraQueryArgs to be overridden for requests.

Reviewed at https://reviews.reviewboard.org/r/6461/